+++ HISTORY.md
@@ -11,6 +11,7 @@
 - session keepalive
 - stream support
 - json5 support
+- context reuse
 
 ## [0.1.0] - 2024-02-xx